2025_ctfzone/web/pied_piper/sources/requirements.txt
flask==3.0.2
flask-sqlalchemy==3.1.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
flask-migrate==4.0.5
gunicorn==21.2.0
openai-whisper
celery==5.5.3
redis==6.2.0
flask-login
requests==2.32.4